distributedembedding-tf/requirements.txt
absl-py
jax[tpu]
tensorflow
tensorflow-datasets
keras=3.10.0
keras-rs==0.2.1
